move NewIPTables to network package; fix license headers
