+++ cmd/shoot_client/app/pathcontroller/pathcontroller_test.go
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Gardener contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
 package pathcontroller
 
 import (
