+++ pkg/ippool/broker_test.go
@@ -1,8 +1,6 @@
-/*
- * SPDX-FileCopyrightText: 2022 SAP SE or an SAP affiliate company and Gardener contributors
- *
- * SPDX-License-Identifier: Apache-2.0
- */
+// SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Gardener contributors
+//
+// SPDX-License-Identifier: Apache-2.0
 
 package ippool
