+++ pkg/network/iptables.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package utils
+package network
 
 import (
 	"fmt"
